Deal with "other" values.
If there are more than 5 values, the lowest values will be grouped together as "Other". This doesn't currently occur in the test data, so you can't see it right now. I manually tested by setting the limit to 3 (instead of 5). The "top 5" categories are determined based on Subject (when available), then Group (when available), then finally overall study, so the y axis can change when going from subject to subject. I could be convinced to go with Study always, but it felt weird to hide the subject's highest category. Maybe save the details for the zoom?