Set up release infrastructure.
Note that this is still intended to merge into `dev`. After this, we should follow its advice and create an `rc` branch to PR into `main`. But also note that R CMD Check freaks out if there isn't a vnum anywhere in news, so I'm prematurely tagging the expected vnum there on purpose.

- Closes #40.